Поправил права с комментариями и обновление картинки объявления

+++ src/main/java/ru/skypro/homework/service/AdsCommentService.java
@@ -4,7 +4,6 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 import ru.skypro.homework.component.Authority;
 import ru.skypro.homework.dto.AdsCommentDto;
 import ru.skypro.homework.dto.ResponseWrapperAdsComment;
@@ -15,9 +14,7 @@ import ru.skypro.homework.mappers.AdsCommentMapper;
 import ru.skypro.homework.repository.AdsRepository;
 import ru.skypro.homework.repository.CommentRepository;
 
-import java.time.Instant;
 import java.util.ArrayList;
-import java.util.Date;
 import java.util.List;
 
 @Service
@@ -33,61 +30,40 @@ public class AdsCommentService {
         this.adsCommentMapper = adsCommentMapper;
     }
 
+    public AdsCommentDto getComment(Integer id) {
+        AdsComment comment = commentRepository.findById(id).get();
+        AdsCommentDto commentDto = adsCommentMapper.toDto(comment);
+        return commentDto;
+    }
+
     public ResponseWrapperAdsComment getAdsComments(Integer adsId) {
         logger.info("getAdsComments invoked");
         ResponseWrapperAdsComment responseWrapperAdsComment;
-
         List<AdsComment> allAdsComments = commentRepository.findByAds_Pk(adsId);
-
-
-
         List<AdsCommentDto> adsCommentDtoList = new ArrayList<>();
-
         allAdsComments.forEach(adsComment -> {
             AdsCommentDto adsCommentDto;
             adsCommentDto = adsCommentMapper.toDto(adsComment);
             adsCommentDtoList.add(adsCommentDto);
         });
         responseWrapperAdsComment = adsCommentMapper.toResponseWrapperAdsComment(adsCommentDtoList);
-
         return responseWrapperAdsComment;
     }
 
-    public AdsCommentDto getComment(Integer adsPk, Integer id) {
-        AdsComment comment = commentRepository.findByPkAndAds_Pk(id, adsPk);
-        AdsCommentDto commentDto = adsCommentMapper.toDto(comment);
-        return commentDto;
-    }
-
-    public void addComments(Integer adsId, AdsCommentDto commentDto) {
-        Date date = Date.from(Instant.now());
+    public void addAdsComment(Integer adsId, AdsCommentDto commentDto, Authentication authentication) {
         AdsComment adsComment = adsCommentMapper.toEntity(commentDto);
-        adsComment.setAuthor(777);
-        adsComment.setCreatedAt(date.toString());
-
+        adsComment.setAuthor(authentication.getName());
         Ads ads = adsRepository.findByPk(adsId);
         adsComment.setAds(ads);
         commentRepository.save(adsComment);
     }
 
-    @Transactional
-    public void deleteComment(Integer id) {
-//        if (Authority.check(authentication).equals(Role.ADMIN.toString()) ||
-//                (Authority.check(authentication).equals(Role.USER.toString())
-//                        && adsRepository.existsByPkAndUser_Username(id, authentication.getName()))) {
-//            commentRepository.deleteById(id);
-//        }
-        commentRepository.deleteById(id);
+    public void deleteAdsComment(Integer id, Authentication authentication) {
+        if (Authority.check(authentication).equals(Role.ADMIN.toString()) ||
+                (Authority.check(authentication).equals(Role.USER.toString())
+                        && commentRepository.existsByPkAndAuthor(id, authentication.getName()))) {
+            commentRepository.deleteById(id);
+        }
     }
 
-
-    @Transactional
-    public AdsCommentDto updateComment(Integer adPk, Integer id, AdsCommentDto commentDto) {
-        logger.info("updateCommentsService");
-        AdsComment adsComment = commentRepository.findByPkAndAds_Pk(id, adPk);
-        adsComment.setText(commentDto.getText());
-        commentRepository.save(adsComment);
-        AdsCommentDto newCommentDto = adsCommentMapper.toDto(adsComment);
-        return newCommentDto;
-    }
 }
